requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==4.2
django-admin-interface==0.30.0
django-colorfield==0.14.0
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-jazzmin==3.0.1
djangorestframework==3.16.0
mysqlclient==2.2.7
pillow==11.2.1
python-slugify==8.0.4
sqlparse==0.5.3
tablib==3.8.0
text-unidecode==1.3
tzdata==2025.2
